pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.63.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
transcribe = "transcribe.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
